backend/requirements.txt
PyMuPDF
transformers
torch
langchain
langchain-community
langchain-huggingface
qdrant-client
python-dotenv
google-generativeai
pillow
pytesseract
fastapi
uvicorn
python-multipart